+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Starting v0.13.0 for Microsoft.Azure.WebJobs.Extensions.OpenAI.Kusto, it will maintain its own [Changelog](./src/WebJobs.Extensions.OpenAI.Kusto/CHANGELOG.md)
+
+Starting v0.1.0 for Microsoft.Azure.WebJobs.Extensions.OpenAI.AzureAISearch, it will maintain its own [Changelog](./src/WebJobs.Extensions.OpenAI)
+
 ## v0.13.0 - Unreleased
 
 ### Added
 
 - Added dotnet-isolated support for Embeddings
-- Added Azure AI Search support for Semantic Search.
+- Added support for multiple search providers.
 - Added dotnet-isolated support for Semantic Search.
 
 ### Breaking changes